pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "attendance-system"
version = "1.0.0"
description = "Intelligent face recognition attendance system with anti-spoofing detection"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Shubham Patra", email = "contact@shubhampatra.com"}
]
keywords = ["attendance", "face-recognition", "computer-vision", "anti-spoofing"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
    "Topic :: Education",
]

# Python 3.9+ for broad compatibility
requires-python = ">=3.9,<4.0"

# Core dependencies with flexible version ranges
dependencies = [
    # Web Framework & Extensions
    "Flask>=3.0,<4.0",
    "Flask-SocketIO>=5.3,<6.0",
    "Flask-RESTX>=1.2,<2.0",
    "Flask-WTF>=1.1,<2.0",
    "Werkzeug>=3.0,<4.0",
    "Jinja2>=3.0,<4.0",
    "eventlet>=0.33,<1.0",
    "python-socketio>=5.10,<6.0",
    "python-engineio>=4.10,<5.0",
    
    # ML / Computer Vision
    "opencv-contrib-python>=4.8,<5.0",
    "insightface>=0.7,<0.8",
    "numpy>=1.24,<2.0",
    "scipy>=1.10,<2.0",
    "Pillow>=10.0,<11.0",
    
    # ONNX Runtime (CPU variant by default, override in requirements/gpu.txt)
    "onnxruntime>=1.17,<2.0",
    
    # Anti-spoofing Runtime
    "torch>=2.0,<3.0",
    "torchvision>=0.15,<1.0",
    
    # Database
    "pymongo>=4.5,<5.0",
    "dnspython>=2.4,<3.0",
    
    # DevOps / Utilities
    "gunicorn>=21.0,<24.0",
    "python-dotenv>=1.0,<2.0",
    
    # Miscellaneous
    "bcrypt>=4.1,<5.0",
    "email-validator>=2.0,<3.0",
    "Markdown>=3.5,<4.0",
    "apscheduler>=3.10,<4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4,<8.0",
    "pytest-cov>=4.1,<5.0",
    "pytest-flask>=1.3,<2.0",
    "pytest-mock>=3.12,<4.0",
    "mongomock>=4.1,<5.0",
    "fakeredis>=2.20,<3.0",
    "black>=23.0,<24.0",
    "flake8>=6.0,<7.0",
    "mypy>=1.5,<2.0",
    "pylint>=2.17,<3.0",
]

gpu = [
    "onnxruntime-gpu>=1.17,<2.0",
]

[project.urls]
Homepage = "https://github.com/ShubhamPatra/attendance_system"
Documentation = "https://github.com/ShubhamPatra/attendance_system#readme"
Repository = "https://github.com/ShubhamPatra/attendance_system"
Issues = "https://github.com/ShubhamPatra/attendance_system/issues"

[tool.setuptools]
packages = [
    "admin_app",
    "anti_spoofing",
    "camera",
    "core",
    "recognition",
    "scripts",
    "student_app",
    "tasks",
    "verification",
    "vision",
]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=. --cov-report=term-missing"
python_files = "test_*.py"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
